+++ config/pos_production.php
@@ -114,6 +114,19 @@ return [
                 'gateway_3d_host' => 'https://vpos.qnbfinansbank.com/Gateway/3DHost.aspx',
             ],
         ],
+        'ziraat-katilim-payfor' => [
+            'name'              => 'ZiraatKatilim-PayFor',
+            'class'             => Mews\Pos\Gateways\PayForPos::class,
+            'gateway_configs'   => [
+                // Ziraat Katilim için hash kontrolü çalışmıyor. O yüzden devre dışı bırakıyoruz.
+                'disable_3d_hash_check' => true,
+            ],
+            'gateway_endpoints' => [
+                'payment_api'     => 'https://vpos.ziraatkatilim.com.tr/Mpi/XMLGate.aspx',
+                'gateway_3d'      => 'https://vpos.ziraatkatilim.com.tr/Mpi/Default.aspx',
+                'gateway_3d_host' => 'https://vpos.ziraatkatilim.com.tr/Mpi/3Dhost.aspx',
+            ],
+        ],
         'vakifbank'            => [
             'name'  => 'VakifBank-VPOS',
             'class' => Mews\Pos\Gateways\PayFlexV4Pos::class,