+++ config/pos_test.php
@@ -65,6 +65,10 @@ return [
         'garanti'              => [
             'name'              => 'Garanti',
             'class'             => Mews\Pos\Gateways\GarantiPos::class,
+            'gateway_configs' => [
+                // GarantiPos'u test ortamda test edebilmek için zorunlu.
+                'test_mode' => true,
+            ],
             'gateway_endpoints' => [
                 'payment_api' => 'https://sanalposprovtest.garantibbva.com.tr/VPServlet',
                 'gateway_3d'  => 'https://sanalposprovtest.garantibbva.com.tr/servlet/gt3dengine',
@@ -77,6 +81,19 @@ return [
                 'payment_api'     => 'https://vpostest.qnbfinansbank.com/Gateway/XMLGate.aspx',
                 'gateway_3d'      => 'https://vpostest.qnbfinansbank.com/Gateway/Default.aspx',
                 'gateway_3d_host' => 'https://vpostest.qnbfinansbank.com/Gateway/3DHost.aspx',
+            ],
+        ],
+        'ziraat-katilim-payfor' => [
+            'name'              => 'ZiraatKatilim-PayFor',
+            'class'             => Mews\Pos\Gateways\PayForPos::class,
+            'gateway_configs'   => [
+                // Ziraat Katilim için hash kontrolü çalışmıyor. O yüzden devre dışı bırakıyoruz.
+                'disable_3d_hash_check' => true,
+            ],
+            'gateway_endpoints' => [
+                'payment_api'     => 'https://payfortestziraatkatilim.cordisnetwork.com/Mpi/XMLGate.aspx',
+                'gateway_3d'      => 'https://payfortestziraatkatilim.cordisnetwork.com/Mpi/Default.aspx',
+                'gateway_3d_host' => 'https://payfortestziraatkatilim.cordisnetwork.com/Mpi/3DHost.aspx',
             ],
         ],
         'vakifbank'            => [
@@ -117,6 +134,10 @@ return [
         'kuveytpos'            => [
             'name'              => 'kuveyt-pos',
             'class'             => Mews\Pos\Gateways\KuveytPos::class,
+            'gateway_configs' => [
+                // testinizi SSL olmayan ortamda yapıyorsanız bu değeri true yapmanız gerekir.
+                'test_mode' => true,
+            ],
             'gateway_endpoints' => [
                 'payment_api' => 'https://boatest.kuveytturk.com.tr/boa.virtualpos.services/Home',
                 'gateway_3d'  => 'https://boatest.kuveytturk.com.tr/boa.virtualpos.services/Home/ThreeDModelPayGate',