+++ examples/finansbank-payfor/_payment_config.php
@@ -16,4 +16,21 @@ $testCards = [
         'name' => 'John Doe',
         'type' => CreditCardInterface::CARD_TYPE_VISA,
     ],
+//    'ziraat-katilim-1' => [
+//        OTP: 123456
+//        'number' => '5352480048848060',
+//        'year' => '31',
+//        'month' => '09',
+//        'cvv' => '131',
+//        'name' => 'John Doe',
+//        'type' => CreditCardInterface::CARD_TYPE_VISA,
+//    ],
+//    'ziraat-katilim-2' => [
+//        'number' => '9792096695823360',
+//        'year' => '29',
+//        'month' => '03',
+//        'cvv' => '123',
+//        'name' => 'John Doe',
+//        'type' => CreditCardInterface::CARD_TYPE_TROY,
+//    ],
 ];