+++ src/Gateways/ParamPos.php
@@ -109,7 +109,10 @@ class ParamPos extends AbstractGateway
             return $this;
         }
 
-        if (!$this->requestDataMapper->getCrypt()->check3DHash($this->account, $request->all())) {
+        if (
+            !$this->is3DHashCheckDisabled()
+            && !$this->requestDataMapper->getCrypt()->check3DHash($this->account, $request->all())
+        ) {
             throw new HashMismatchException();
         }
 
@@ -159,7 +162,10 @@ class ParamPos extends AbstractGateway
      */
     public function make3DPayPayment(Request $request, array $order, string $txType): PosInterface
     {
-        if (!$this->requestDataMapper->getCrypt()->check3DHash($this->account, $request->request->all())) {
+        if (
+            !$this->is3DHashCheckDisabled()
+            && !$this->requestDataMapper->getCrypt()->check3DHash($this->account, $request->request->all())
+        ) {
             throw new HashMismatchException();
         }
 
@@ -173,7 +179,10 @@ class ParamPos extends AbstractGateway
      */
     public function make3DHostPayment(Request $request, array $order, string $txType): PosInterface
     {
-        if (!$this->requestDataMapper->getCrypt()->check3DHash($this->account, $request->request->all())) {
+        if (
+            !$this->is3DHashCheckDisabled()
+            && !$this->requestDataMapper->getCrypt()->check3DHash($this->account, $request->request->all())
+        ) {
             throw new HashMismatchException();
         }