+++ tests/Unit/Gateways/EstPosTest.php
@@ -10,6 +10,7 @@ use Mews\Pos\Client\HttpClient;
 use Mews\Pos\Crypt\CryptInterface;
 use Mews\Pos\DataMapper\RequestDataMapper\RequestDataMapperInterface;
 use Mews\Pos\DataMapper\ResponseDataMapper\ResponseDataMapperInterface;
+use Mews\Pos\Entity\Account\AbstractPosAccount;
 use Mews\Pos\Entity\Account\EstPosAccount;
 use Mews\Pos\Entity\Card\CreditCardInterface;
 use Mews\Pos\Event\RequestDataPreparedEvent;
@@ -113,18 +114,7 @@ class EstPosTest extends TestCase
             ->method('getCrypt')
             ->willReturn($this->cryptMock);
 
-        $this->pos = new EstPos(
-            $this->config,
-            $this->account,
-            $this->requestMapperMock,
-            $this->responseMapperMock,
-            $this->serializerMock,
-            $this->eventDispatcherMock,
-            $this->httpClientMock,
-            $this->loggerMock,
-        );
-
-        $this->pos->setTestMode(true);
+        $this->pos = $this->createGateway($this->config);
 
         $this->card = CreditCardFactory::createForGateway(
             $this->pos,
@@ -134,6 +124,20 @@ class EstPosTest extends TestCase
             '122',
             'ahmet',
             CreditCardInterface::CARD_TYPE_VISA
+        );
+    }
+
+    private function createGateway(array $config, ?AbstractPosAccount $account = null): PosInterface
+    {
+        return new EstPos(
+            $config,
+            $account ?? $this->account,
+            $this->requestMapperMock,
+            $this->responseMapperMock,
+            $this->serializerMock,
+            $this->eventDispatcherMock,
+            $this->httpClientMock,
+            $this->loggerMock,
         );
     }
 
@@ -148,6 +152,7 @@ class EstPosTest extends TestCase
         $this->assertSame([PosInterface::CURRENCY_TRY], $this->pos->getCurrencies());
         $this->assertSame($this->config, $this->pos->getConfig());
         $this->assertSame($this->account, $this->pos->getAccount());
+        $this->assertFalse($this->pos->isTestMode());
     }
 
     /**
@@ -227,6 +232,40 @@ class EstPosTest extends TestCase
         $this->assertTrue($pos->isSuccess());
     }
 
+    /**
+     * @return void
+     */
+    public function testMake3DHostPaymentWithoutHashCheckSuccess(): void
+    {
+        $config = $this->config;
+        $config += [
+            'gateway_configs' => [
+                'disable_3d_hash_check' => true,
+            ],
+        ];
+
+        $pos = $this->createGateway($config);
+
+        $this->cryptMock->expects(self::never())
+            ->method('check3DHash');
+
+        $testData = EstPosResponseDataMapperTest::threeDHostPaymentDataProvider()['success1'];
+        $request  = Request::create('', 'POST', $testData['paymentData']);
+        $order    = $testData['order'];
+        $txType   = $testData['txType'];
+
+        $this->responseMapperMock->expects(self::once())
+            ->method('map3DHostResponseData')
+            ->with($request->request->all(), $txType, $order)
+            ->willReturn($testData['expectedData']);
+
+        $pos->make3DHostPayment($request, $order, $txType);
+
+        $result = $pos->getResponse();
+        $this->assertSame($testData['expectedData'], $result);
+        $this->assertTrue($pos->isSuccess());
+    }
+
     public function testMake3DHostPaymentHashMismatchException(): void
     {
         $data = EstPosResponseDataMapperTest::threeDHostPaymentDataProvider()['success1']['paymentData'];
@@ -264,6 +303,40 @@ class EstPosTest extends TestCase
             ->willReturn($testData['expectedData']);
 
         $pos = $this->pos;
+
+        $pos->make3DPayPayment($request, $order, $txType);
+
+        $result = $pos->getResponse();
+        $this->assertSame($testData['expectedData'], $result);
+        $this->assertTrue($pos->isSuccess());
+    }
+
+    /**
+     * @return void
+     */
+    public function testMake3DPayPaymentWithoutHashCheckSuccess(): void
+    {
+        $config = $this->config;
+        $config += [
+            'gateway_configs' => [
+                'disable_3d_hash_check' => true,
+            ],
+        ];
+
+        $pos = $this->createGateway($config);
+
+        $this->cryptMock->expects(self::never())
+            ->method('check3DHash');
+
+        $testData = EstPosResponseDataMapperTest::threeDPayPaymentDataProvider()['success1'];
+        $request  = Request::create('', 'POST', $testData['paymentData']);
+        $order    = $testData['order'];
+        $txType   = $testData['txType'];
+
+        $this->responseMapperMock->expects(self::once())
+            ->method('map3DPayResponseData')
+            ->with($request->request->all(), $txType, $order)
+            ->willReturn($testData['expectedData']);
 
         $pos->make3DPayPayment($request, $order, $txType);
 
@@ -522,6 +595,86 @@ class EstPosTest extends TestCase
         $this->assertSame($isSuccess, $this->pos->isSuccess());
     }
 
+    /**
+     * @dataProvider make3DPaymentWithoutHashCheckDataProvider
+     */
+    public function testMake3DPaymentWithoutHashCheck(
+        array   $order,
+        string  $txType,
+        Request $request,
+        array   $paymentResponse,
+        array   $expectedResponse,
+        bool    $is3DSuccess,
+        bool    $isSuccess
+    ): void {
+        $config = $this->config;
+        $config += [
+            'gateway_configs' => [
+                'disable_3d_hash_check' => true,
+            ],
+        ];
+
+        $pos = $this->createGateway($config);
+
+        $this->cryptMock->expects(self::never())
+            ->method('check3DHash');
+
+        $this->responseMapperMock->expects(self::once())
+            ->method('extractMdStatus')
+            ->with($request->request->all())
+            ->willReturn('3d-status');
+
+        $this->responseMapperMock->expects(self::once())
+            ->method('is3dAuthSuccess')
+            ->with('3d-status')
+            ->willReturn($is3DSuccess);
+
+        $create3DPaymentRequestData = [
+            'create3DPaymentRequestData',
+        ];
+        if ($is3DSuccess) {
+            $this->requestMapperMock->expects(self::once())
+                ->method('create3DPaymentRequestData')
+                ->with($this->account, $order, $txType, $request->request->all())
+                ->willReturn($create3DPaymentRequestData);
+
+            $this->configureClientResponse(
+                $txType,
+                'https://entegrasyon.asseco-see.com.tr/fim/api',
+                $create3DPaymentRequestData,
+                'request-body',
+                'response-body',
+                $paymentResponse,
+                $order,
+                PosInterface::MODEL_3D_SECURE
+            );
+
+            $this->responseMapperMock->expects(self::once())
+                ->method('map3DPaymentData')
+                ->with($request->request->all(), $paymentResponse, $txType, $order)
+                ->willReturn($expectedResponse);
+        } else {
+            $this->responseMapperMock->expects(self::once())
+                ->method('map3DPaymentData')
+                ->with($request->request->all(), null, $txType, $order)
+                ->willReturn($expectedResponse);
+            $this->requestMapperMock->expects(self::never())
+                ->method('create3DPaymentRequestData');
+            $this->serializerMock->expects(self::never())
+                ->method('encode');
+            $this->serializerMock->expects(self::never())
+                ->method('decode');
+            $this->eventDispatcherMock->expects(self::never())
+                ->method('dispatch');
+        }
+
+        $pos->make3DPayment($request, $order, $txType);
+
+        $result = $pos->getResponse();
+        $this->assertSame($expectedResponse, $result);
+        $this->assertSame($isSuccess, $pos->isSuccess());
+    }
+
     public function testMake3DPaymentHashMismatchException(): void
     {
         $data = EstPosResponseDataMapperTest::threeDPaymentDataProvider()['3d_auth_success_payment_fail']['threeDResponseData'];
@@ -685,6 +838,38 @@ class EstPosTest extends TestCase
                 'is3DSuccess'     => false,
                 'isSuccess'       => false,
             ],
+            '3d_auth_success_payment_fail' => [
+                'order'           => EstPosResponseDataMapperTest::threeDPaymentDataProvider()['3d_auth_success_payment_fail']['order'],
+                'txType'          => EstPosResponseDataMapperTest::threeDPaymentDataProvider()['3d_auth_success_payment_fail']['txType'],
+                'request'         => Request::create(
+                    '',
+                    'POST',
+                    EstPosResponseDataMapperTest::threeDPaymentDataProvider()['3d_auth_success_payment_fail']['threeDResponseData']
+                ),
+                'paymentResponse' => EstPosResponseDataMapperTest::threeDPaymentDataProvider()['3d_auth_success_payment_fail']['paymentData'],
+                'expected'        => EstPosResponseDataMapperTest::threeDPaymentDataProvider()['3d_auth_success_payment_fail']['expectedData'],
+                'is3DSuccess'     => true,
+                'isSuccess'       => false,
+            ],
+            'success'                      => [
+                'order'           => EstPosResponseDataMapperTest::threeDPaymentDataProvider()['success1']['order'],
+                'txType'          => EstPosResponseDataMapperTest::threeDPaymentDataProvider()['success1']['txType'],
+                'request'         => Request::create(
+                    '',
+                    'POST',
+                    EstPosResponseDataMapperTest::threeDPaymentDataProvider()['success1']['threeDResponseData']
+                ),
+                'paymentResponse' => EstPosResponseDataMapperTest::threeDPaymentDataProvider()['success1']['paymentData'],
+                'expected'        => EstPosResponseDataMapperTest::threeDPaymentDataProvider()['success1']['expectedData'],
+                'is3DSuccess'     => true,
+                'isSuccess'       => true,
+            ],
+        ];
+    }
+
+    public static function make3DPaymentWithoutHashCheckDataProvider(): array
+    {
+        return [
             '3d_auth_success_payment_fail' => [
                 'order'           => EstPosResponseDataMapperTest::threeDPaymentDataProvider()['3d_auth_success_payment_fail']['order'],
                 'txType'          => EstPosResponseDataMapperTest::threeDPaymentDataProvider()['3d_auth_success_payment_fail']['txType'],