+++ tests/Unit/Gateways/VakifKatilimTest.php
@@ -126,8 +126,6 @@ class VakifKatilimTest extends TestCase
             $this->loggerMock,
         );
 
-        $this->pos->setTestMode(true);
-
         $this->card = CreditCardFactory::createForGateway(
             $this->pos,
             '4155650100416111',
@@ -150,6 +148,7 @@ class VakifKatilimTest extends TestCase
         $this->assertSame([PosInterface::CURRENCY_TRY], $this->pos->getCurrencies());
         $this->assertSame($this->config, $this->pos->getConfig());
         $this->assertSame($this->account, $this->pos->getAccount());
+        $this->assertFalse($this->pos->isTestMode());
     }
 
     /**